requirements.txt
# Core UI / Backend
streamlit
Flask==3.1.2
pandas
plotly
requests
python-dotenv

# Tezos
# NOTE: pytezos requires the native libsodium library. Install it with:
#   macOS:  brew install libsodium
#   Ubuntu: sudo apt-get install libsodium-dev
pytezos
smartpy-tezos

# Ethereum
web3
eth_account
py-solc-x

# Solana / Anchor
anchorpy==0.18.0
anchorpy-core==0.1.3
solana==0.30.2
solders==0.18.1
base58==2.1.1
based58==0.1.1
nest-asyncio==1.6.0
toml==0.10.2
